Rename to pepper sync

+++ zingolib/src/wallet/data.rs
@@ -482,9 +482,9 @@ pub mod summaries {
 
     use chrono::DateTime;
     use json::JsonValue;
+    use pepper_sync::keys::KeyId;
     use zcash_primitives::{consensus::BlockHeight, transaction::TxId};
     use zingo_status::confirmation_status::ConfirmationStatus;
-    use zingo_sync::keys::KeyId;
 
     use crate::{
         error::BuildError,
@@ -1771,7 +1771,7 @@ pub mod summaries {
     /// Not to be used for internal logic in the system.
     #[derive(Clone, PartialEq, Debug)]
     pub struct OutgoingNoteSummary {
-        pub output_index: usize,
+        pub output_index: u16,
         pub key_id: KeyId,
         pub value: u64,
         pub memo: Option<String>,
